Keep the CPS pass from creating proc points due to unreachable parents.
The parser/flattener will generate an extra block after an if/else 
statement even if both branches exit the function.  So it is possible
for the input to the CPS pass to have dead/unreachable blocks.

If a dead block goes to a live block then the live block would
have more parents than the dead block and prior to this patch
the live block would then be identified as a proc point.
This is fixed by adding a check to see if the parent
has at least one owner.