Fix a knot-tying bug with ghc --make
This bug showed up when I touched Data.Generics.Schemes, and then said
	make build.library.base
This compiles the base library with --make, and I got an assertion failure
in TcIface.  The explanation is in Note [Tying the knot] in TcIface.

As well as fixing the bug, I also impoved the assertion message.

This might fix or improve other knot-tying problems with --make