Fix Trac #1525
A Name used to have a Parent, but no longer has.  When we want to 
print info about data type T and data constructor MkT, the info about
T already describes MkT so we want to discard the latter.  Now that
Names don't have a Parent, we must do that in a different way,
using implicitTyThings

Test is ghci011