In GHCi, filter instances by what is in scope, not just by what is in scope unqualified
Trac #1581 was doing too much filtering; it even filtered out intances 
defined in this very module!  The new rule shows more instances, but
hopefully not to many.

Furthermore I have moved the filtering out of TcRnDriver (where it does
not belong) to InteractiveEval. And I've added a note to the documentation.