fix of wanted equational class context
Previously failed to account for equational
class context for wanted dictionary contraints, e.g. wanted C a
in 

	class a ~ Int => C a
	instance C Int

should give rise to wanted a ~ Int and consequently discharge a ~ Int by
unifying a with Int and then discharge C Int with the instance.

All ancestor equalities are taken into account.