Change spill cost function back to inverse length of live range.
On further testing it turns out that using Chaitin's spill cost formula
of counting the number of uses of a var and then dividing by the degree
of the node isn't actually a win. Perhaps because this isn't counting 
the number of actual spills inserted in the output code.

This would be worth revisiting if other work manages to increase the 
register pressure in the native code.