Add -dasm-lint
When -dasm-lint is turned on the register conflict graph is checked for 
internal consistency after each build/color pass. Make sure that all 
edges point to valid nodes, that nodes are colored differently to their
neighbours, and if the graph is supposed to be colored, that all nodes
actually have a color.