Fix bug #1725 (Haddock links between packages) Resolving this bug is a bit tricky, it boils down to the question: Should the Haddock links between packages include the package version or not?
Pro: We can differentiate between various versions of the same package,
installed all at once. (How often does this really happen in practice?)

Cons: When package A refers to a package B, and B is later upgraded, links
in A's documentation will break. Furthermore, if an *additional* version of
B is installed, which version should A refer to?

Because IMHO it is not clear what to do when version numbers are included,
let's leave them out. If somebody has a better idea, feel free to submit a
better patch.

MERGE TO STABLE