In a pattern binding, a type sig in the pattern cannot bind a type variable
In a pattern binding with a pattern type signature, such as

	(Just (x::a)) = e

the pattern type signature cannot bind type variables.  Hence
'a' must be in scope already for the above example to be legal.

This has been the situation for some time, but Dan changed it when
adding view patterns.  This one-line change restores the old behaviour.