FIX Trac #1825: standalone deriving Typeable
Standalone deriving of typeable now requires you to say
	instance Typeable1 Maybe
which is exactly the shape of instance decl that is generated
by a 'deriving( Typeable )' clause on the data type decl.

This is a bit horrid, but it's the only consistent way, at least
for now.  If you say something else, the error messages are helpful.

MERGE to 6.8 branch