Update test output
This is actually *slightly* worse than before, because of the
error-message refactoring in TcUnify. The previous message 
reported a mis-match between
    (a->Bool) and (Bool -> Bool)
whereas the new one only complains of mismatch between
    a   and   Bool
Trouble is, the old one was vulnerable to the problem
described in the comments on tcSubExp. 

So I'm accepting this regression for now; let's see if
anyone notices!