Improve inlining for INLINE non-functions
(No need to merge to 6.8, but no harm if a subsequent patch needs it.)

The proximate cause for this patch is to improve the inlining for INLINE
things that are not functions; this came up in the NDP project.  See
Note [Lone variables] in CoreUnfold.

This caused some refactoring that actually made things simpler.  In 
particular, more of the inlining logic has moved from SimplUtils to 
CoreUnfold, where it belongs.