Do not worker/wrapper INLINE things, even if they are in a recursive group
This patch stops the worker/wrapper transform working on an INLINE thing,
even if it's in a recursive group.  It might not be the loop breaker.  Indeed
a recursive group might have no loop breaker, if the only recursion is 
through rules.

Again, this change was provoked by one of Roman's NDP libraries.
Specifically the Rec { splitD, splitJoinD } group in 
	Data.Array.Parallel.Unlifted.Distributed.Arrays

Simon