FIX #2049, another problem with the module context on :reload The previous attempt to fix this (#1873, #1360) left a problem that occurred when the first :load of the program failed (#2049).
Now I've implemented a different strategy: between :loads, we remember
all the :module commands, and just replay them after a :reload.  This
is in addition to remembering all the package modules added with
:module, which is orthogonal.

This approach is simpler than the previous one, and seems to do the
right thing in all the cases I could think of.  Let's hope this is the
last bug in this series...